Apko keyrings to recognize key suffixes when lost during fetch: (#1630)
When download keys from Artifactory, such URLs for e.g. are being used:


https://artifactory.{redacted}.com/artifactory/api/security/keypair/public/repositories/development-wolfi-virtual

The URL returns:
```
   content-disposition: attachment; filename=keyname.rsa.pub ,
```

The keyname can be configured to be e.g. 'development-wolfi-virtual'.

Before this patch there was a mismatch on INDEX files signatures
verification.
The verification expected: development-wolfi-virtual.rsa.pub,
but as Apko ignores the content-disposition header, it observes just
`development-wolfi-virtual.rsa.pub`
that does not match.

In this PR I enabled matching that recognized development-wolfi-virtual
~= development-wolfi-virtual.rsa.pub.

Alternatively we could consider:
a) changing fetcher to recognize content-disposition -> but this becomes
tricky to handle with --offline or caching modes.
b) when we fetch the keys... always assume they have .rsa.pub extension
-> and add it. In theory it could be a breaking change
(if there are valid ?? scenarios for keys without .rsa.pub extension).

---------

Signed-off-by: Piotr Tabor (ptab) <132376425+sfc-gh-ptabor@users.noreply.github.com>