Drop cosign dep (#1645)
The dependency graph of cosign isn't suitable for use as a library.

It may be worth revisiting this to see if it's feasible to compose apko
and cosign at the command line in a nice way, but since we aren't using
this feature of apko ourselves, I'd like to drop it to improve compile
times for everything downstream of us.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>