Reimplement Test_GenerateCacheFile() using mock data
Test_GenerateCacheFile() operated on the developer's system directly,
so we left it disabled by default. It wasn't a very useful test
in itself because it only generated a cache file - it didn't check
its contents.

Let's instead create a cache file from a testdata directory and
read it back in to verify its contents. Instead of ELF binaries,
our libraries are actually YAML files, populated by a mock implementation
of getElfInfo().

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>