usrmerge: apko needs to create apk dirs in /usr/lib instead of /lib (#1593)
- There's a bootstrapping issue where apko was creating /lib before
wolfi-baselayout was trying symlink /lib to /usr/lib in a package build
via melange. Melange is using apko to setup it's build environment which
would create the /lib directory. Upon trying to install the usrmerge'd
wolfi-baselayout, it's symlink from /lib -> /usr/lib could not be
created because the directory already existed and everything fails
afterwards.
- Because melange uses apko and not apk, this works for our builds.
However, apk would try to do the same if its db was reinitialized.

Try to deal with all possible situations where /lib is one of:
1. An existing directory.
2. Nothing at all yet.
3. A symlink to usr/lib
4. Exists with an existing apk directory (therefore database).