Make apko hermetic again. (#1654)
Apko started to ignore lock files, i.e. even if the lock_file to `apko
build` is specified explicitly, the configuration resolution was
calling: `LockImageConfiguration() -> BuildPackageList() ->
ResolveWorld()` thus
depending on the current (up to date) state of the `ApkIndex`.

This PR makes sure that list of packages defined in the lock file takes
precedence (blocks) reevaluation of the list of packages to install.