Reverse sort entries to match ldconfig
I spent hours trying to figure out why I was getting an error when starting
an application using cache generated by this package, but not by glibc's
ldconfig. I discovered that ldconfig's entries do not match the order
of the strings they point to in the string table. The entries appear to be
sorted reverse-alphabetically by the stringtable string they point to.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>