Proof of concept for lazily compressed layers
I want to speed up some interactions with mounting filesystems into
melange. The way we do it today is dumb because we gzip and then
immediately gunzip.

Downside of this approach is that we need to store the uncompressed
filesystem somehow as well. One idea to mitigate that would be removing
the uncompressed tar if the caller ever asks for the compressed stream
so that we don't keep both?

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>