lock: fix up filter for wanted arches
Currently when a multiarch lock file is passed in, and only a single
arch is requested to be built, the call to Arch2LockedPackages returns
all arches present in the lockfile, instead of only the desired
arches, which triggers creation of multi arch layers.

```
$ apko build ./apko.yaml wolfi-base:latest wolfi-base.tar --lockfile apko.lock.json  --arch amd64  --log-level debug
...
2025/05/12 22:12:21 DEBU Building images for 1 architectures: [amd64]
...
xnox says pls = l.Arch2LockedPackages(input.Archs) returned 2 for inputs 1
2025/05/12 22:12:21 DEBU loading config file: ./apko.yaml
2025/05/12 22:12:21 DEBU loading config file: ./apko.yaml
2025/05/12 22:12:21 DEBU doing pre-flight checks arch=x86_64
2025/05/12 22:12:21 DEBU initializing apk database arch=x86_64
2025/05/12 22:12:21 DEBU finished initializing apk database arch=x86_64
2025/05/12 22:12:21 DEBU doing pre-flight checks arch=aarch64
2025/05/12 22:12:21 DEBU initializing apk database arch=aarch64
```

I can't quite understand the nested if ok blocks. Instead going with a
more pythonic solution. Parse the lockfile as is, then iterate desired
arches, copy the ones we like, and return them. It uses more memory,
but is less lines of code with less nesting, and is still linear in
time.

This partially reverts:
- https://github.com/chainguard-dev/apko/pull/1664