ldsocache snapshot/20250507 (#1665)
This catches us up with my current tree:
https://github.com/dannf/ldso-cache/commits/devel/. The really important
bit here is
https://github.com/chainguard-dev/apko/pull/1665/commits/d6b89ebf202d8176f11113846c0ad705dbaa21da,
which fixes a bug that can cause libraries to not be found at runtime
even though they are in the cache because `ld` apparently requires them
to be sorted.

It also does away with an optimization I had where it would not add a
library to the cache if it had already added a library with the same
name from a different directory. I had done this to replicate
`ldconfig`'s behavior, but I misunderstood that behavior. `ldconfig`
does add multiple libs of the same name to the cache. But it does seem
to detect when 2 libs are actually the same file (e.g. /lib/libfoo.so.4
and /usr/lib/libfoo.so.4 when /lib -> /usr/lib). I think a better answer
for us would be to just stop listing symlinked dirs in
`/etc/ld.so.conf.d/*` snippets.

Also, this finally adds testing for the file scanning logic using a mock
elf parser.

---------

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>