sbom: avoid accidental pruning of entries
Ensure that SBOM entries which have an identical 'name' field don't
get pruned from the Image SBOM; 'name' does not need to be unique so
use 'version' as well to ensure that the correct target element is
selected.

Simplify the replacement of the package with the Images ID; the match
on the name is not unique and is actually surplus to requirements as
the target elements are now selected by package name and version.

Add a test case based on unbound to validate.

Fixes #1507

Signed-off-by: James Page <james.page@chainguard.dev>