sbom: resolve issues with missing/invalid image SBOM information (#1672)
Prune a whole load of code out of the processing of APK SBOM's into the
Image SBOM; this ensures that the SBOM for the Image is complete and we
don't accidentally prune packages with the same name (which is not
guaranteed to be unique).

I'm sure there is history here and I'm not 100% sure I understand why a
generated SPDXRef was pushed to the Image SBOM document, only to be
removed at the end of the APK SBOM processing. Testing looks good based
on comparing current APKO release with this PR.

Add a test case based on unbound to validate.

Fixes #1507

Signed-off-by: James Page <james.page@chainguard.dev>

---------

Signed-off-by: James Page <james.page@chainguard.dev>