Add support for the =~ constraint (#1681)
We were trying to use the `=~` constraint via the Apko Terraform
Provider which should be
[valid](https://wiki.alpinelinux.org/wiki/Alpine_Package_Keeper#Package_pinning):
```sh
/ # apk add yara-x=~0.14.0
(1/1) Installing yara-x (0.14.0-r3)
Executing busybox-1.37.0-r40.trigger
OK: 166 MiB in 16 packages
```
but this was not a valid `matcher` case so we defaulted to `versionAny`
which would instead pull in the latest version.

This PR adds a case for the `=~` constraint and treats it like `~` so
that we correctly resolve versions via `includesVersion` and updates a
few tests to validate both the `~` and `=~` constraints.

FWIW it does look like `~` works as expected.

---------

Signed-off-by: egibs <20933572+egibs@users.noreply.github.com>