Write a partial installed db to every layer (#1752)
Some scanners operate on a per-layer basis, so having our installed
database only exist in the top-most layer leads to false positives.

By including the portion of the installed db that is relevant to each
layer, we can appease those scanners at the cost of a small amount of
duplication in each layer. I don't love this, but it's an option.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>