Add support for runtime-only repositories (#1786)
apko currently supports "build_repositories" (used at build-time) and
"repositories" (used a build-time + run-time). This adds support for
"runtime_repositories" which are not used for building the image but are
only installed into the `/etc/apk/repositories` file of the resulting
image.

The current `repositories` field was named `runtimeRepositories` in many
places through the Go code. For clarity, I renamed this to
`repositories` to avoid confusion with this new field. This change is
non-breaking with the json/yaml schemas since this field was already
serialized using the `repositories` field name.

I see a couple references to `build.WithExtraRuntimeRepos` outside of
this repo, so I've left a backwards-compatible implementation in place.
I plan to move that usage to `build.WithExtraRepos` to free up the
`build.WithExtraRuntimeRepos` name for runtime-only repos.