Add support for runtime-only repositories (#1790)
This is largely the same change as
https://github.com/chainguard-dev/apko/pull/1786, which we reverted.
apko currently supports "build_repositories" (used at build-time) and
"repositories" (used a build-time + run-time). This adds support for
"runtime_repositories" which are not used for building the image but are
only installed into the /etc/apk/repositories file of the resulting
image.

This change is backwards compatible for users of `apko` as a CLI tool
(since we aren't renaming anything in the serialized file format), but
it is breaking for users of `apko` as a Go library. The existing go
struct field "RuntimeRepositories" is renamed to "Repositories" to avoid
confusion with this new configuration option. The new field is called
"RuntimeOnlyRepositories" rather than "RuntimeRepositories" so that go
code importing apko as a library doesn't silently switch to this new field
when bumping the version.