Simplify the sorting of paths for apk db/installed file.
The change here allows us to just use sort.Slice on []tar.Headers.
It also makes it much easier to test, as we can simply test the
comparison of 2 different entries and assume that sort.Slice is correct.

The thing that is still here is that my changes are not dropping
intermediate paths. There is a test in TestSortTarHeaders named
intermediate_dirs_in_the_tree_should_be_required_to_preserve_children
that is failing now.

That is addressed by the following commits.