Revert "fix: include root-level files in the installed apk database. (#1801)"
This reverts commit 37187b1a56a69b50af35c52d822b017d7fece984.

The generated /lib/dpk/db/installed files were not readable by apk.
apk would fail with messages like this:

    / # apk info -L
    ERROR: FDB format error (line 16, entry 'R')
    ERROR: Unable to read database state: database file format error
    ERROR: Failed to open apk database: database file format error

The problem is that apk requires 'R:' entries to have a preceeding
'F' (directory) entry.  If the only entries in / are directories,
then this is not a problem.  In order to put an 'R:' entry in /,
it must be preceeded by a 'F:' (a directory entry with "" as value).