Reduce the log noise from auto-auth failures.
Today virtually all of our elastic builds surface errors because of this, but it's just noise.

This changes things to hide the chainctl error, reduce the severity of the (current) warning, and document how a user can reproduce the auth failure, if this is why things are not working for them.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>