bug-fix: handles usrmerge base image correctly (#1856)
When building on a base image with usrmerge layout, we should not try to
create the `lib` directory, because the provided packages already has
it.

To avoid making this decision based on actually looking at the base
image's file system, we will just look at the install packages to see if
it was one that provides a usrmerge layout.

Also add a "examples test" to run the sample's shell script to confirm
fix.

Signed-off-by: Nghia Tran <tcnghia@gmail.com>