Add support for tagged repositories (#1868)
## Problem

Using tagged repositories users can then also have control from which
repository to install the package.

### This works as expected and is all fine

Currently an apko.yaml config file allows for the following:

```yaml
contents:
  keyring:
    - https://packages.wolfi.dev/os/wolfi-signing.rsa.pub
  repositories:
    - https://dl-cdn.alpinelinux.org/alpine/v3.22/main
    - https://dl-cdn.alpinelinux.org/alpine/v3.22/community
    - '@edgecommunity https://dl-cdn.alpinelinux.org/alpine/edge/community'
    - '@edgetesting https://dl-cdn.alpinelinux.org/alpine/edge/testing'
  packages:
    - alpine-baselayout
    - alpine-keys
    - apk-tools
    - aws-cli=~2.31@edgecommunity
    - tenv@edgetesting
```

Using above config all packages will be pulled from the v3.22
repositories unless tagged. This allows to select for specific packages
a more recent version.

This allows for adding tagged repositories. See
https://wiki.alpinelinux.org/wiki/Repositories#Tagged_repository

However doing this in a programmed manner it isn't supported as the
validation is too strict.

### This fails and is fixed by this PR

```go
imgCfg := types.ImageConfiguration{
		Contents: types.ImageContents{
			Repositories: []{"@edgecommunity https://dl-cdn.alpinelinux.org/alpine/edge/community",
		},
	}
```

## Solution

This is where this PR comes in, it enhances the repository validation to
allow tagging. It also adds unit-tests for the repository validation.