apk: partially fix offline cached builds
Even in offline case, calls to releases.json were performed without
caching. And caching them, appears to not work (possibly bug in
cachePathFromURL() as there are not enough sub-directories).

However, offline cache (even if not fully working) is needed to avoid
looping trying to download releases.json when one doesn't have
internet and is building with --offline flag.

Also downgrade all Warnings around not able to fetch keys to Debug
level, as cache for them is yet fully working, and thus it is
red-herrings for the user that otherwise provides keys via keyring
options.

The below should work:

1. apko build examples/alpine-base.yaml foo bar
2. cut Wi-Fi
3. apko build --ignore-signatures --offline examples/alpine-base.yaml foo bar

If/when caching of key discovery is fixed, `--ignore-signatures`
shouldn't be needed to build offline.