apk: fix offline cached builds with alpine key discovery (#1938)
Even in offline case, calls to releases.json were performed without
caching. And caching them, appears to not work (possibly bug in
cachePathFromURL() as there are not enough sub-directories).

Offline cache is needed to avoid looping trying to download
releases.json when one doesn't have
internet and is building with --offline flag.

Then cache discovery needed to be fixed up, as for the first time there
is a possibility of a subdirectory in the cached dir. (thanks to
@javacruft)

Also downgrade all Warnings around not able to fetch keys to Debug
level, as most of the time they are red-herring (alpine URLs do not have
Chainguard key discovery, and most artifactory urls do not have
Chainguard key discovery either potentially).