Gracefully handle cached files being deleted (#1959)
The global cache in the `apk` package stores cached
`expandapk.APKExpanded` structs in memory, which references files on
disk and contains readers for those files. The in-memory data is
currently never invalidated, which can lead to problems if the files on
disk are deleted. This PR updates the cache to check if the backing file
still exists before returning data from the cache. If it's absent, the
APK data is re-populated.