Add the ability to specify additional certs to be added to the image (#1977)
This allows users to specify certificates that apko should be appending
to the default truststores that are shipped via packages (via
ca-certificates-bundle for example). The additional certs are specified
via a new `certificates` key that looks like this

```
certificates:
  additional:
    - name: my-cert
      content: |
        some extra content about my certificate that's useful to me as a user
        -----BEGIN CERTIFICATE-----
        ...
        -----END CERTIFICATE-----
    - name: my-other-cert
      content: |
        -----BEGIN CERTIFICATE-----
        ...
        -----END CERTIFICATE-----
```

Each content section must include exactly one certificate. Name and
content are required fields.

apko also installs the individual certificats to
/usr/local/share/ca-certificates so that they're not lost if
`update-ca-certificates` is run.