Check for validity of the cached expanded APK more thoroughly
Instead of just checking for the presence of a single file, this ensures that all active pieces of the expandapk structure are actually still valid. It checks that all necessary files exist and checks that cached file handles are actually the ones pointing to the files present in the filesystem.