Add support for SBOMs without shortcut fields.
SPDX considers documentDescribes as a shortcut field - it is not present
in SPDX Go types (even though it will unmarshal them correctly).
However, if apko is presented an SBOM without a shortcut field, it will fail
to extract out the package details, even though the SBOM is considered valid by
SPDX standards.

This adds support to look for the equivalent details in the DESCRIBES
relationship.

See https://github.com/spdx/tools-golang/pull/201 for more context.