cli/dot: Link nodes to apk.dag.dev
This commit adds a link to the APK registry viewer at apk.dag.dev.  But
the user can override that with a flag to point to a registry explorer
of their choosing which is most suitable for the dependencies that they
are reviewing.  This replaces the preexisting link function that linked
within the SVG to specific nodes.