cli/dot: Rendering improvements (#1992)
- **cli/dot: Link nodes to apk.dag.dev**
This commit adds a link to the APK registry viewer at apk.dag.dev. But
  the user can override that with a flag to point to a registry explorer
of their choosing which is most suitable for the dependencies that they
are reviewing. This replaces the preexisting link function that linked
  within the SVG to specific nodes.
  

- **cli/dot: Add descriptions as tooltips**
Here I add some code to set the node's tooltip to the description of the
  package.  Some packages do not have descriptions, such as virtual
  packages.
  

- **cli/dot: Remove constant "images" tooltip**
  

- **cli/dot: Add edge tooltips**
  Here I add a tooltip for the graph edges, this will indicate the two
  nodes that this edge connects and the kind of dependency that it is
  representing.  This should be beneficial when looking at rather
  complicated graphs.