Detect broken symlinks in the cache and fix them
Currently, when an external process deletes a file that a symlink is pointing to in the cache, but not the symlink itself, we will never fix that and the cache will be permanently broken.

This makes the `AdvertiseCachedFile` sensitive to the possibility of links being broken. Broken links are removed and the advertisement process is retried to establish a functioning cache state again.